requirements.txt
#Django==4.2.6
django_phonenumber_field==7.2.0
django-phonenumber-field[phonenumbers]
django-cors-headers
#run...> pip install -r requirements.txt <...to install all the required packages